+++ tests/Unit/Service/Client/ServiceClientTest.php
@@ -25,6 +25,7 @@ namespace OAT\Library\Lti1p3Core\Tests\Unit\Service\Client;
 use Cache\Adapter\PHPArray\ArrayCachePool;
 use Carbon\Carbon;
 use GuzzleHttp\ClientInterface;
+use GuzzleHttp\Exception\ClientException;
 use OAT\Library\Lti1p3Core\Exception\LtiException;
 use OAT\Library\Lti1p3Core\Registration\RegistrationInterface;
 use OAT\Library\Lti1p3Core\Security\Jwt\Builder\Builder;
@@ -36,6 +37,7 @@ use OAT\Library\Lti1p3Core\Tests\Traits\NetworkTestingTrait;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 
 class ServiceClientTest extends TestCase
 {
@@ -99,7 +101,8 @@ class ServiceClientTest extends TestCase
                     'GET',
                     'http://example.com',
                     [
-                        'headers' => ['Authorization' => 'Bearer access_token']
+                        'headers' => ['Authorization' => 'Bearer access_token'],
+                        'http_errors' => true
                     ]
                 ]
             )
@@ -140,7 +143,8 @@ class ServiceClientTest extends TestCase
                     'GET',
                     'http://example.com',
                     [
-                        'headers' => ['Authorization' => 'Bearer access_token']
+                        'headers' => ['Authorization' => 'Bearer access_token'],
+                        'http_errors' => true
                     ]
                 ]
             )
@@ -172,7 +176,8 @@ class ServiceClientTest extends TestCase
                 'GET',
                 'http://example.com',
                 [
-                    'headers' => ['Authorization' => 'Bearer cached_access_token']
+                    'headers' => ['Authorization' => 'Bearer cached_access_token'],
+                    'http_errors' => true
                 ]
             )
             ->willReturn(
@@ -200,7 +205,8 @@ class ServiceClientTest extends TestCase
                 'GET',
                 'http://example.com',
                 [
-                    'headers' => ['Authorization' => 'Bearer cached_access_token']
+                    'headers' => ['Authorization' => 'Bearer cached_access_token'],
+                    'http_errors' => true
                 ]
             )
             ->willReturn(
@@ -211,6 +217,68 @@ class ServiceClientTest extends TestCase
 
         $this->assertInstanceOf(ResponseInterface::class, $result);
         $this->assertEquals('service response', $result->getBody()->__toString());
+    }
+
+    public function testItCanPerformAServiceCallFromPopulatedInvalidTokenCacheAndAutoRetry(): void
+    {
+        $scopes = ['scope1', 'scope2'];
+
+        $cacheKey = $this->generateAccessTokenCacheKey($this->registration, $scopes);
+        $cacheItem = $this->cache->getItem($cacheKey)->set('invalid_access_token');
+        $this->cache->save($cacheItem);
+
+        $this->clientMock
+            ->expects($this->exactly(3))
+            ->method('request')
+            ->withConsecutive(
+                [
+                    'GET',
+                    'http://example.com',
+                    [
+                        'headers' => ['Authorization' => 'Bearer invalid_access_token'],
+                        'http_errors' => true
+                    ]
+                ],
+                [
+                    'POST',
+                    $this->registration->getPlatform()->getOAuth2AccessTokenUrl(),
+                    [
+                        'form_params' => [
+                            'grant_type' => ServiceClientInterface::GRANT_TYPE,
+                            'client_assertion_type' => ClientAssertionCredentialsGrant::CLIENT_ASSERTION_TYPE,
+                            'client_assertion' => $this->createTestClientAssertion($this->registration),
+                            'scope' => 'scope1 scope2'
+                        ]
+                    ]
+                ],
+                [
+                    'GET',
+                    'http://example.com',
+                    [
+                        'headers' => ['Authorization' => 'Bearer valid_access_token'],
+                        'http_errors' => true
+                    ]
+                ]
+            )
+            ->willReturnOnConsecutiveCalls(
+                $this->throwException(
+                    new ClientException(
+                        'invalid token',
+                        $this->createMock(ServerRequestInterface::class),
+                        $this->createResponse('invalid token', 401)
+                    )
+                ),
+                $this->createResponse(json_encode(['access_token'=> 'valid_access_token', 'expires_in' => 3600]), 201),
+                $this->createResponse('service response')
+            );
+
+        $result = $this->subject->request($this->registration, 'GET', 'http://example.com', [], $scopes);
+
+        $this->assertInstanceOf(ResponseInterface::class, $result);
+        $this->assertEquals('service response', $result->getBody()->__toString());
+
+        $this->assertTrue($this->cache->hasItem($cacheKey));
+        $this->assertEquals('valid_access_token', $this->cache->getItem($cacheKey)->get());
     }
 
     public function testItThrowAnLtiExceptionOnMissingToolKeyChain(): void
@@ -304,6 +372,44 @@ class ServiceClientTest extends TestCase
         $this->subject->request($this->registration, 'GET', 'http://example.com');
     }
 
+    public function testItThrowAnLtiExceptionOnInvalidPlatformEndpointResponse(): void
+    {
+        $this->expectException(LtiException::class);
+        $this->expectExceptionMessage('Cannot perform request');
+
+        $this->clientMock
+            ->expects($this->exactly(2))
+            ->method('request')
+            ->withConsecutive(
+                [
+                    'POST',
+                    $this->registration->getPlatform()->getOAuth2AccessTokenUrl(),
+                    [
+                        'form_params' => [
+                            'grant_type' => ServiceClientInterface::GRANT_TYPE,
+                            'client_assertion_type' => ClientAssertionCredentialsGrant::CLIENT_ASSERTION_TYPE,
+                            'client_assertion' => $this->createTestClientAssertion($this->registration),
+                            'scope' => ''
+                        ]
+                    ]
+                ],
+                [
+                    'GET',
+                    'http://example.com',
+                    [
+                        'headers' => ['Authorization' => 'Bearer access_token'],
+                        'http_errors' => true
+                    ]
+                ]
+            )
+            ->willReturnOnConsecutiveCalls(
+                $this->createResponse(json_encode(['access_token'=> 'access_token', 'expires_in' => 3600])),
+                'invalid response'
+            );
+
+        $this->subject->request($this->registration, 'GET', 'http://example.com');
+    }
+
     public function testItThrowAnLtiExceptionOnPlatformEndpointFailure(): void
     {
         $this->expectException(LtiException::class);
@@ -329,16 +435,88 @@ class ServiceClientTest extends TestCase
                     'GET',
                     'http://example.com',
                     [
-                        'headers' => ['Authorization' => 'Bearer access_token']
+                        'headers' => ['Authorization' => 'Bearer access_token'],
+                        'http_errors' => true
                     ]
                 ]
             )
             ->willReturnOnConsecutiveCalls(
                 $this->createResponse(json_encode(['access_token'=> 'access_token', 'expires_in' => 3600])),
-                'invalid output'
+                $this->throwException(
+                    new ClientException(
+                        'internal server error',
+                        $this->createMock(ServerRequestInterface::class),
+                        $this->createResponse('internal server error', 500)
+                    )
+                )
             );
 
         $this->subject->request($this->registration, 'GET', 'http://example.com');
+    }
+
+    public function testItThrowAnLtiExceptionOnPlatformEndpointFailureAfterAutoRetry(): void
+    {
+        $this->expectException(LtiException::class);
+        $this->expectExceptionMessage('Cannot perform request: internal server error after retry');
+
+        $scopes = ['scope1', 'scope2'];
+
+        $cacheKey = $this->generateAccessTokenCacheKey($this->registration, $scopes);
+        $cacheItem = $this->cache->getItem($cacheKey)->set('invalid_access_token');
+        $this->cache->save($cacheItem);
+
+        $this->clientMock
+            ->expects($this->exactly(3))
+            ->method('request')
+            ->withConsecutive(
+                [
+                    'GET',
+                    'http://example.com',
+                    [
+                        'headers' => ['Authorization' => 'Bearer invalid_access_token'],
+                        'http_errors' => true
+                    ]
+                ],
+                [
+                    'POST',
+                    $this->registration->getPlatform()->getOAuth2AccessTokenUrl(),
+                    [
+                        'form_params' => [
+                            'grant_type' => ServiceClientInterface::GRANT_TYPE,
+                            'client_assertion_type' => ClientAssertionCredentialsGrant::CLIENT_ASSERTION_TYPE,
+                            'client_assertion' => $this->createTestClientAssertion($this->registration),
+                            'scope' => 'scope1 scope2'
+                        ]
+                    ]
+                ],
+                [
+                    'GET',
+                    'http://example.com',
+                    [
+                        'headers' => ['Authorization' => 'Bearer valid_access_token'],
+                        'http_errors' => true
+                    ]
+                ]
+            )
+            ->willReturnOnConsecutiveCalls(
+                $this->throwException(
+                    new ClientException(
+                        'invalid token',
+                        $this->createMock(ServerRequestInterface::class),
+                        $this->createResponse('invalid token', 401)
+                    )
+                ),
+                $this->createResponse(json_encode(['access_token'=> 'valid_access_token', 'expires_in' => 3600]), 201),
+                $this->throwException(
+                    new ClientException(
+                        'internal server error after retry',
+                        $this->createMock(ServerRequestInterface::class),
+                        $this->createResponse('internal server error after retry', 500)
+                    )
+                )
+            );
+
+        $this->subject->request($this->registration, 'GET', 'http://example.com', [], $scopes);
     }
 
     private function generateAccessTokenCacheKey(RegistrationInterface $registration, array $scopes = []): string